server/requirements.txt
Flask==2.1.1
bcrypt==3.2.0
pycryptodome==3.10.1
pyotp==2.6.0
pymongo==3.12.1
